Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    float stupeflip;
-    scanf("%f",&stupeflip);
-    printf("You entered: %.4f",stupeflip);
+    double stupeflip;
+    scanf("%d",&stupeflip);
+    printf("You entered: %.4d",stupeflip);
     return 0;
 }
